Fixed bug with frontend logoutting
