+++ frontend/src/containers/Cart/Cart.js
@@ -8,4 +8,4 @@ const Cart = () => {
   );
 };
 
-export default Cart;
+export default Cart
